Use the gzip library on the server.  Probably doesn't speed things 	up as currently implemented, but does avoid hassles in terms of 	finding an external gzip program. 	* zlib.c, server.h (gunzip_and_write, read_and_gzip): Now returns 	whether a fatal error occurred, rather than expecting error (1, 	...) to work. 	* client.c (update_entries, send_modified): Change callers. 	* server.c (receive_file): Rewrite gzip code to use 	gunzip_and_write rather than filter_through_gunzip. 	(server_updated): Likewise, use read_and_gzip rather than 	filter_through_gzip. 	* client.c, client.h (filter_through_gzip, filter_through_gunzip), 	run.c, cvs.h (filter_stream_through_program): Removed; no longer used. 	* sanity.sh (server): New tests server-4 and server-5 test 	this feature (note that CVS 1.10 also passes these tests; the 	behavior is supposed to be unchanged).
CVS-ID: src/ChangeLog 1.1649
CVS-ID: src/client.c 1.262
CVS-ID: src/client.h 1.35
CVS-ID: src/cvs.h 1.194
CVS-ID: src/run.c 1.27
CVS-ID: src/sanity.sh 1.463
CVS-ID: src/server.c 1.208
CVS-ID: src/server.h 1.22
CVS-ID: src/zlib.c 1.7

+++ src/sanity.sh
@@ -17708,7 +17708,52 @@ EOF
 	    fi
 	    dotest server-3 "test -d ${TESTDIR}/crerepos/CVSROOT" ""
 
+	    # Now some tests of gzip-file-contents (used by jCVS).
+	    awk 'BEGIN {
+printf "%c%c%c%c%c%c.6%c%c+I-.%c%c%c%c5%c;%c%c%c%c", 
+31, 139, 8, 0, 5, 7, 0, 3, 225, 2, 0, 198, 185, 5, 0, 0, 0}' \
+	      </dev/null >gzipped.dat
+	    echo Root ${TESTDIR}/crerepos >session.dat
+	    # Note that the CVS client sends "-b 1.1.1", and this
+	    # test doesn't.  But the server also defaults to that.
+	    cat <<EOF >>session.dat
+UseUnchanged
+gzip-file-contents 3
+Argument -m
+Argument msg
+Argumentx 
+Argument dir1
+Argument tag1
+Argument tag2
+Directory .
+in-real-life-this-is-funky-but-server-seems-to-ignore-it
+Modified file1
+u=rw,g=r,o=r
+z25
+EOF
+	    cat gzipped.dat >>session.dat
+	    echo import >>session.dat
+	    if ${testcvs} server >${TESTDIR}/server.tmp <session.dat; then
+	      dotest server-4 "cat ${TESTDIR}/server.tmp" "M N dir1/file1
+M 
+M No conflicts created by this import
+M 
+ok"
+	    else
+	      echo "exit status was $?" >>${LOGFILE}
+	      fail server-4
+	    fi
+	    dotest server-5 \
+"${testcvs} -q -d ${TESTDIR}/crerepos co -p dir1/file1" "test"
+
+	    if test "$keep" = yes; then
+	      echo Keeping ${TESTDIR} and exiting due to --keep
+	      exit 0
+	    fi
+
 	    rm -rf ${TESTDIR}/crerepos
+	    rm gzipped.dat session.dat
+	    rm ${TESTDIR}/server.tmp
 	  fi # skip the whole thing for local
 	  ;;
